requirements.txt
tensorflow>=2.15.0
pillow
opencv-python
streamlit
numpy
